requirements.txt
torch
torchvision
transformers
gradio
pillow
scikit-learn
numpy
openpyxl